ml_db_project/requirements.txt
SQLAlchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
requests==2.32.4
gpxpy==1.6.2
python-dateutil==2.9.0.post0
tqdm==4.67.1
pandas
numpy
scikit-learn
statsmodels
imbalanced-learn
joblib
scipy
folium
ydata-profiling
sweetviz
requests
Pillow
streamlit==1.41.1
plotly==5.24.1
streamlit-autorefresh==1.0.1
networkx==3.4.2
statsmodels==0.14.6